28 fixe form validation

+++ assets/js/order-confirmation.js
@@ -3,7 +3,12 @@ window.addEventListener("DOMContentLoaded", (event) => {
     categories = ["teddies", "cameras", "furniture"];
 
     getOrderData().then(data => {
-        console.log(data);
+
+        if(data.orderIds.length <= 0) {
+            window.location.href = "/";
+            return;
+        }
+
         renderOrderData(data);
 
         localStorage.clear();
@@ -18,15 +23,15 @@ function getOrderData() {
             let orderIds = [];
             let totalPrice = 0;
 
-            categories.forEach(category => {
-                let id = localStorage.getItem("order-id-confirmation-" + category);
-                if(typeof(id) == "undefined") {
-                    reject();
-                    return;
-                }
+            for(var i = 0; i < categories.length; i++) {
+                let category = categories[i];
 
-                orderIds.push(id);
-            })
+                let id = localStorage.getItem("order-id-confirmation-" + category);
+                if(id == null)
+                    continue;
+
+                orderIds.push(getCategoryName(category) + ": " + id);
+            }
 
             totalPrice = parseFloat(localStorage.getItem("order-price-confirmation"));
 
@@ -53,4 +58,28 @@ function renderOrderData(data) {
 function toEuro(number) {
     number = number/100;
     return new Intl.NumberFormat('fr-FR', { style: 'currency', currency: 'EUR' }).format(number);
+}
+
+function getCategoryName(type) {
+    let str = "";
+
+    switch(type) {
+        case "teddies":
+            str = "Ours en peluche";
+        break;
+
+        case "cameras":
+            str = "Caméras";
+        break;
+
+        case "furniture":
+            str = "Meubles";
+        break;
+
+        default:
+            str = "Unknown";
+        break;
+    }
+
+    return str;
 }
